ingestion/requirements.txt
python-dotenv==1.0.0
qdrant-client==1.8.0
google-generativeai==0.4.0
tiktoken==0.5.2